D-14518 break down avatar into respective parts, fix lint rules, and ensure doc site doesn't improperly style avatar imgs
