+++ packages/components/src/Avatar/LetterAvatar.js
@@ -1,0 +1,13 @@
+import PropTypes from 'prop-types';
+
+export const LetterAvatar = ({ text }) => {
+  const firstLetter = text.length > 0 ? text[0] : '';
+  return firstLetter;
+};
+
+LetterAvatar.propTypes = {
+  /**
+   * Title of the avatar
+   */
+  text: PropTypes.string.isRequired,
+};
